+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Foraging period files can contain multiple years of data (#60)
 - Patches have coordinates for visualization; calculated if not provided (#65)
 - Daily patch visits for pollen and nectar are counted, adds respective observers (#66)
+- Stats for foraging rounds are recorded, adds respective table observer (#67)
 
 ### Bugfixes